refactor: move OpenAI-compatible base to private module
