+++ pydantic_ai_slim/pydantic_ai/providers/bedrock_mantle.py
@@ -12,7 +12,7 @@ from pydantic_ai.profiles import merge_profile
 from pydantic_ai.profiles.openai import OpenAIModelProfile, openai_model_profile
 from pydantic_ai.providers._bedrock_model_names import split_bedrock_model_id
 
-from .openai import _OpenAICompatibleProvider  # pyright: ignore[reportPrivateUsage]
+from ._openai_compatible import OpenAICompatibleProvider as _OpenAICompatibleProvider
 
 try:
     from openai import AsyncBedrockOpenAI, AsyncOpenAI
